requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.10
notebook==7.0.6
ipykernel==6.28.0
ipywidgets==8.1.1

# ML Frameworks (pre-installed for convenience)
torch==2.1.2
tensorflow==2.15.0
jax[cpu]==0.4.23
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Database (optional)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Caching (optional)
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
